# Conflicts:
#   backend/src/main/java/com/ssafy/tedbear/global/common/oauth2/OAuth2LoginSuccessHandler.java
#   backend/src/main/java/com/ssafy/tedbear/global/common/oauth2/jwt/JwtProvider.java

+++ backend/src/main/java/com/ssafy/tedbear/domain/sentence/service/SentenceBookmarkService.java
@@ -1,6 +1,7 @@
 package com.ssafy.tedbear.domain.sentence.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 
@@ -14,6 +15,7 @@ import com.ssafy.tedbear.domain.sentence.dto.SentenceBookmarkStatusDto;
 import com.ssafy.tedbear.domain.sentence.entity.Sentence;
 import com.ssafy.tedbear.domain.sentence.entity.SentenceBookmark;
 import com.ssafy.tedbear.domain.sentence.repository.SentenceBookmarkRepository;
+import com.ssafy.tedbear.global.common.FindMemberService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,27 +26,46 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class SentenceBookmarkService {
 	private final SentenceBookmarkRepository sentenceBookmarkRepository;
+	private final FindMemberService findMemberService;
 
-	public void saveSentenceBookmark(Member member, SentenceBookmarkDto sentenceBookmarkDto) {
-		SentenceBookmark sentenceBookmark = sentenceBookmarkDto.toEntity(member);
+	public void saveSentenceBookmark(String memberUid, SentenceBookmarkDto sentenceBookmarkDto) {
+		Member member = findMemberService.findMember(memberUid);
+		Sentence sentence = new Sentence(sentenceBookmarkDto.getSentenceNo());
+		if (isPresent(member, sentence)) {
+			throw new IllegalArgumentException("이미 북마크된 문장입니다.");
+		}
+		SentenceBookmark sentenceBookmark = sentenceBookmarkDto.toEntity(member, sentence);
 		sentenceBookmarkRepository.save(sentenceBookmark);
 	}
 
-	public void deleteSentenceBookmark(Member member, SentenceBookmarkDto sentenceBookmarkDto) {
-		Sentence sentence = sentenceBookmarkDto.toSentenceEntity();
+	public void deleteSentenceBookmark(String memberUid, SentenceBookmarkDto sentenceBookmarkDto) {
+		Member member = findMemberService.findMember(memberUid);
+		Sentence sentence = new Sentence(sentenceBookmarkDto.getSentenceNo());
+		if (!isPresent(member, sentence)) {
+			throw new IllegalArgumentException("해당 문장의 북마크 정보가 존재하지 않습니다.");
+		}
 		sentenceBookmarkRepository.deleteByMemberAndSentence(member, sentence);
 	}
 
-	public SentenceBookmarkStatusDto getBookmarkStatus(Member member, Long sentenceId) {
+	public SentenceBookmarkStatusDto getBookmarkStatus(String memberUid, Long sentenceId) {
+		Member member = findMemberService.findMember(memberUid);
 		Sentence sentence = Sentence.builder().no(sentenceId).build();
 		boolean isBookmarked = sentenceBookmarkRepository.findByMemberAndSentence(member, sentence).isPresent();
 		return new SentenceBookmarkStatusDto(isBookmarked);
 	}
 
-	public SentenceBookmarkDetailDto.ListResponse getBookmarkList(Long memberId, Pageable pageable) {
-		List<SentenceBookmarkDetailDto> bookmarkedList = sentenceBookmarkRepository.findSentenceByMember(memberId,
-			pageable).getContent();
-		return new SentenceBookmarkDetailDto.ListResponse(bookmarkedList);
+	public SentenceBookmarkDetailDto.ListResponse getBookmarkList(String memberUid, Pageable pageable) {
+		Member member = findMemberService.findMember(memberUid);
+		List<Sentence> bookmarkedSenteceList = sentenceBookmarkRepository.findSentenceByMember(member, pageable)
+			.getContent();
+
+		return new SentenceBookmarkDetailDto.ListResponse(bookmarkedSenteceList);
+	}
+
+	public boolean isPresent(Member member, Sentence sentence) {
+		Optional<SentenceBookmark> byMemberAndSentence = sentenceBookmarkRepository.findByMemberAndSentence(member,
+			sentence);
+		return byMemberAndSentence.isPresent();
 	}
 
 }
